+++ mypyc/irbuild/classdef.py
@@ -11,7 +11,7 @@ from mypyc.ops import (
     object_rprimitive, bool_rprimitive, dict_rprimitive, is_optional_type, is_object_rprimitive,
     is_none_rprimitive,
 )
-from mypyc.ops_misc import (
+from mypyc.primitives.misc_ops import (
     dataclass_sleight_of_hand, py_setattr_op, pytype_from_template_op, py_calc_meta_op,
     type_object_op, py_hasattr_op, not_implemented_op, true_op
 )
