+++ mypyc/irbuild/expression.py
@@ -19,7 +19,7 @@ from mypyc.ops import (
     object_rprimitive, is_none_rprimitive, FUNC_CLASSMETHOD, FUNC_STATICMETHOD
 )
 from mypyc.ops_primitive import name_ref_ops
-from mypyc.ops_misc import new_slice_op, iter_op, ellipsis_op, type_op
+from mypyc.primitives.misc_ops import new_slice_op, iter_op, ellipsis_op, type_op
 from mypyc.ops_list import new_list_op, list_append_op, list_extend_op
 from mypyc.ops_tuple import list_tuple_op
 from mypyc.primitives.dict_ops import new_dict_op, dict_set_item_op
