+++ mypyc/irbuild/for_helpers.py
@@ -15,7 +15,7 @@ from mypyc.ops import (
 )
 from mypyc.primitives.int_ops import unsafe_short_add
 from mypyc.ops_list import list_len_op, list_get_item_unsafe_op
-from mypyc.ops_misc import iter_op, next_op
+from mypyc.primitives.misc_ops import iter_op, next_op
 from mypyc.ops_exc import no_err_occurred_op
 
 if TYPE_CHECKING:
