+++ mypyc/irbuild/specialize.py
@@ -21,7 +21,7 @@ from mypyc.ops import (
     Value, RType, RTuple, BasicBlock, LoadInt, RaiseStandardError, Unreachable, OpDescription,
     str_rprimitive, list_rprimitive, dict_rprimitive, set_rprimitive, bool_rprimitive
 )
-from mypyc.ops_misc import true_op, false_op
+from mypyc.primitives.misc_ops import true_op, false_op
 from mypyc.irbuild.builder import IRBuilder
