+++ mypyc/ops_primitive.py
@@ -222,4 +222,4 @@ import mypyc.ops_str  # noqa
 import mypyc.ops_list  # noqa
 import mypyc.primitives.dict_ops  # noqa
 import mypyc.ops_tuple  # noqa
-import mypyc.ops_misc  # noqa
+import mypyc.primitives.misc_ops  # noqa
