+++ mypyc/primitives/misc_ops.py
